requirements.txt
asyncio==3.4.3
boto3==1.28.57
botocore==1.31.57
celery==5.3.0
deepgram-sdk==2.11.0
Django==3.2.21
django-storages==1.14
djangorestframework==3.14.0
drf-extra-fields==3.7.0
ffmpeg-python==0.2.0
gunicorn==21.2.0
imageio==2.31.2
imageio-ffmpeg==0.4.9

Pillow==9.5.0
psycopg2-binary==2.9.8
python-decouple==3.8
redis==5.0.1
regex==2023.8.8
whitenoise==6.